Move CassandraQueryNormalizer to cassandra instrumentations
